+++ module/product/src/Infrastructure/Calculator/TranslationInheritanceCalculator.php
@@ -8,13 +8,13 @@ declare(strict_types=1);
 
 namespace Ergonode\Product\Infrastructure\Calculator;
 
-use Ergonode\Attribute\Domain\Entity\AbstractAttribute;
 use Ergonode\Core\Domain\Query\LanguageQueryInterface;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Value\Domain\ValueObject\StringCollectionValue;
 use Ergonode\Value\Domain\ValueObject\StringValue;
 use Ergonode\Value\Domain\ValueObject\TranslatableStringValue;
 use Ergonode\Value\Domain\ValueObject\ValueInterface;
+use Ergonode\Attribute\Domain\ValueObject\AttributeScope;
 
 class TranslationInheritanceCalculator
 {
@@ -28,7 +28,7 @@ class TranslationInheritanceCalculator
     /**
      * @return string|array|null
      */
-    public function calculate(AbstractAttribute $attribute, ValueInterface $value, Language $language)
+    public function calculate(AttributeScope $scope, ValueInterface $value, Language $language)
     {
         $languagesPath = $this->languageQuery->getInheritancePath($language);
         $calculatedValue = null;
@@ -36,7 +36,7 @@ class TranslationInheritanceCalculator
             $translations = $value->getValue();
             $find = false;
             $setUp = false;
-            if ($attribute->getScope()->isLocal()) {
+            if ($scope->isLocal()) {
                 foreach ($languagesPath as $inheritance) {
                     if ($inheritance->isEqual($language)) {
                         $find = true;
@@ -50,7 +50,7 @@ class TranslationInheritanceCalculator
                     }
                 }
             }
-            if ($attribute->getScope()->isGlobal()) {
+            if ($scope->isGlobal()) {
                 $inheritance = $this->languageQuery->getRootLanguage();
                 if (array_key_exists($inheritance->getCode(), $translations)) {
                     $calculatedValue = $translations[$inheritance->getCode()];